server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
openai>=1.50.0
pydantic>=2.11.7
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1